requirements.txt
greenlet==1.1.3
Pillow==9.2.0
playwright==1.26.1
pyee==8.1.0
typing-extensions==4.3.0
websockets==10.1
